DialogAgentJson now a base class
