+++ src/main/scala/org/clulab/asist/DialogAgentStdin.scala
@@ -14,7 +14,7 @@ import java.util.Scanner
 
 class DialogAgentStdin (
     override val nMatches: Int = 0
-) extends AgentJson { 
+) extends DialogAgentJson { 
 
   override val source_type: String = ""
   println("\nDialog Agent stdin extractor running.")
